+++ 5.0/.gitignore
@@ -1,0 +1,6 @@
+build
+node_modules
+yarn.lock
+.idea
+.vscode
+.nvmrc
